requirements.txt
# arabic-reshaper
asgiref     
asttokens==2.0.5      
Babel==2.9.1
beautifulsoup4==4.10.0
cairocffi==1.1.0      
CairoSVG==2.7.0      
certifi==2022.12.7    
cffi==1.14.3
chardet==3.0.4        
Click==7.0
colorama==0.4.4       
coreapi
coreschema==0.0.4     
coverage==4.0.3       
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
django==3.1.14
django-appconf==1.0.5
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-import-export==2.4.0
django-jazzmin==2.2.9
django-safedelete==0.5.6
django-softdelete
django-taggit==1.4.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-generators==0.5.0
easy-password-generator==0.0.7
et-xmlfile==1.0.1
executing==0.8.2
exponent-server-sdk==1.0.1
Faker==4.1.3
fpdf==1.7.2
future==0.18.3
html5lib==1.1
icecream==2.1.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
num2words==0.5.10
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==9.3.0
pip-tools==4.4.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==0.10.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-bidi==0.4.2
python-dateutil==2.8.1
qrcode==5.1
reportlab
sentry-sdk==1.14.0
six==1.14.0
soft-delete==0.2.2
soupsieve==2.3
sqlparse==0.4.4
tablib==2.0.0
text-unidecode==1.3
tinycss2==1.0.2
uritemplate==3.0.1
urllib3
pytest_mock
WeasyPrint==52.5
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
psycopg2-binary
gunicorn==20.0.4
django-filter==21.1
pusher
drf-api-tracking==1.8.0
cryptography==41.0.0
stripe==2.63.0
python-dotenv==0.21.0
deep-translator
paydunya
volkeno-paydunya-gateway-py==1.0.0
google-currency==1.0.10
# SpeechRecognition
geopy==2.3.0
django-phone-field==1.8.1
django-resized==1.0.2
drf-yasg==1.21.7